requirements.txt
numpy=1.18.2
opencv-python=4.2.0
chumpy=0.69
open3d-python=0.7.0
matplotlib=3.1.1
torch=1.4
torchvision=0.5.0
scikit-learn
tqdm
